+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 9.3.0 (git master)
 * Improve Task critical flag calculation when reading PMXML files.
+* Improve support for Synchro 6.3 SP files.
 
 ## 9.2.0 (30/03/2021)
 * Improve accuracy when reading subprojects from MPP files.